style : 반응형 및 애니메이션 적용 fix : MyPage 탭 로딩 이슈로 인한 컴포넌트 분리 진행중
