+++ src/components/MyPage/PlanPreview.jsx
@@ -25,11 +25,11 @@ export default function PlanPreview({ userSchedule }) {
 
   if (userSchedule) {
     return (
-      <div className="swiper-slide mx-auto flex flex-col gap-4 lg:w-[768px] xl:w-[1236px]">
+      <div className=" mx-auto flex w-auto flex-col  gap-4 xl:w-[1236px]">
         {userSchedule?.map((item) => (
           <Link to={`/myschedule/${item.id}`} key={item.id}>
             <div
-              className=" animate-fade-slide relative mx-auto flex  cursor-pointer justify-center 
+              className="  relative mx-auto flex  cursor-pointer justify-center 
              rounded-xl  border-2 
             bg-[hsl(0,0%,90%,0.6)] bg-fixed opacity-80 transition duration-300 ease-in-out hover:opacity-100 hover:contrast-125"
             >
