+++ src/pages/DetailPage.jsx
@@ -42,6 +42,7 @@ export default function DetailPage() {
 
   // 쿼리 키
   const queryKey = ['recommends', recommendId];
+  console.log(queryKey);
 
   // 데이터 쿼리
   const { isLoading, error, data } = useQuery({
@@ -117,6 +118,8 @@ export default function DetailPage() {
 
     // 로그인 사용자 북마크 여부
     const isBookmark = detailPlace.userEmail.includes(user.id);
+    // console.log(detailPlace);
+    // console.log(isBookmark);
 
     // 북마크 토글 함수
     const handleToggleBookmark = (recommendId, userId) => async () => {
@@ -125,39 +128,41 @@ export default function DetailPage() {
     };
 
     return (
-      <div className=" mx-auto  min-h-screen min-w-[22.5rem] bg-background pb-10">
+      <div className="   mx-auto min-h-screen min-w-[22.5rem] bg-background pb-10">
         <div className="container">
           <div className=" flex flex-col items-center justify-center gap-10 ">
-            <h1 className="sr-only">추천 장소</h1>
             <MyPageHeader page="detail" />
+            <h1 className="sr-only">추천 장소</h1>
+          </div>
+          <section className="animate-fade-animate px-6">
             <img
               src={getPocketHostImageURL(detailPlace, 'image')}
               alt={`${detailPlace.place} 이미지`}
-              className="  mx-3 h-[22.5rem] min-h-[22.5rem] w-[77.25rem] min-w-[20rem] rounded-xl object-cover"
+              className="  mx-auto my-10 h-[22.5rem] min-h-[22.5rem] w-[77.25rem] min-w-[20rem] rounded-xl object-cover "
             />
-          </div>
-          <div className="mx-auto  my-5 max-w-[1236px] px-4">
-            <span className=" h-4 rounded-full bg-secondary px-2 py-1 text-[0.875rem]">
-              {detailPlace.localMain}
-            </span>
-            <div className="mb-1  mt-3 flex items-center">
-              <h2 className=" text-2xl font-bold text-contentsPrimary">
-                {detailPlace.place}
-              </h2>
-              <button
-                type="button"
-                onClick={handleToggleBookmark(detailPlace.id, user.id)}
-                className="m-2"
-              >
-                <BookMark color={isBookmark ? '#C9ECFF' : ''} />
-              </button>
+            <div className="mx-auto  my-5 max-w-[1236px] px-4">
+              <span className=" h-4 rounded-full bg-secondary px-2 py-1 text-[0.875rem]">
+                {detailPlace.localMain}
+              </span>
+              <div className="mb-1  mt-3 flex items-center">
+                <h2 className=" text-2xl font-bold text-contentsPrimary">
+                  {detailPlace.place}
+                </h2>
+                <button
+                  type="button"
+                  onClick={handleToggleBookmark(detailPlace.id, user.id)}
+                  className="m-2"
+                >
+                  <BookMark color={isBookmark ? '#C9ECFF' : ''} />
+                </button>
+              </div>
+              <p className="mb-3 text-[0.875rem] font-light text-gray-1">
+                {detailPlace.address}
+              </p>
+              <div className="border-b-[0.0625rem] border-contentsPrimary"></div>
             </div>
-            <p className="mb-3 text-[0.875rem] font-light text-gray-1">
-              {detailPlace.address}
-            </p>
-            <div className="border-b-[0.0625rem] border-contentsPrimary"></div>
-          </div>
-          <DetailInfo detailPlace={detailPlace} />
+            <DetailInfo detailPlace={detailPlace} />
+          </section>
         </div>
       </div>
     );
